test_agents/pyproject.toml
[project]
name = "test-agents"
version = "0.1.0"
description = "Test agents for A2A Platform"
requires-python = ">=3.11"
dependencies = [
    "google-adk[a2a]>=0.1.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "litellm>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["math_agent", "text_agent"]
